Switch from float to double and write to /dev/null

+++ capnp/thing.capnp
@@ -1,7 +1,7 @@
 @0xb3676cb33073faa5;
 
 struct MyThing {
-    thing @0 :Float32;
+    thing @0 :Float64;
 }
 
 struct AllThings {
